Added plaid api information updated dependencies
