+++ requirements.txt
@@ -7,3 +7,4 @@ sqlalchemy-serializer==1.4.1
 flask==3.0.2
 python-dotenv==1.0.1
 openai==1.13.3
+plaid-python==19.0.0
